refactor(provider): remove passport read model from application code
