--- backend/src/main/java/com/alligator/market/backend/provider/infrastructure/startup/passport/projection/ProviderPassportProjectionStartupRunner.java
@@ -1,34 +1,0 @@
-package com.alligator.market.backend.provider.infrastructure.startup.passport.projection;
-
-import com.alligator.market.backend.provider.readmodel.passport.projection.service.ProviderPassportProjectionService;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.ApplicationArguments;
-import org.springframework.boot.ApplicationRunner;
-
-import java.util.Objects;
-
-/**
- * Запуск проекции паспортов провайдеров при старте приложения.
- */
-@Slf4j
-public final class ProviderPassportProjectionStartupRunner implements ApplicationRunner {
-
-    private final ProviderPassportProjectionService service;
-
-    public ProviderPassportProjectionStartupRunner(ProviderPassportProjectionService service) {
-        this.service = Objects.requireNonNull(service, "service must not be null");
-    }
-
-    @Override
-    public void run(ApplicationArguments args) {
-        log.info("Provider passport projection started");
-
-        try {
-            service.project();
-            log.info("Provider passport projection finished");
-        } catch (RuntimeException ex) {
-            log.error("Provider passport projection failed", ex);
-            throw ex; // fail-fast: срываем старт приложения
-        }
-    }
-}
